begin gummyroll refactor
