+++ contracts/programs/gummyroll/Cargo.toml
@@ -18,3 +18,4 @@ default = []
 [dependencies]
 anchor-lang = { path="../../../deps/anchor/lang" }
 bytemuck = "1.8.0"
+concurrent-merkle-tree = { path="../../../lib/concurrent-merkle-tree" }
