+++ lib/concurrent-merkle-tree/Cargo.toml
@@ -1,0 +1,11 @@
+[workspace]
+[package]
+name = "concurrent-merkle-tree"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+solana-program = "1.10.10"
+bytemuck = "1.8.0"
+# rand_distr = "0.4.3"
+# rand = "0.7"
